requirements.txt
youtube_dl
python-dotenv